# src/main/resources/openapi/swagger.yaml
swagger: "2.0"
info:
  description: "API for managing fishermen"
  version: "1.0.0"
  title: "Fisherman API"
basePath: "/fishermansHelp"
schemes:
  - "https"
consumes:
  - "application/json"
produces:
  - "application/json"
paths:
  /fishermen:
    post:
      tags:
        - "Fisherman Management"
      summary: "Add new fisherman profile"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - in: "body"
          name: "body"
          description: "Fisherman object that needs to be added to the system"
          required: true
          schema:
            $ref: "#/definitions/Fisherman"
      responses:
        "201":
          description: "Fisherman created"
        "400":
          description: "Invalid input"
    get:
      tags:
        - "Fisherman Management"
      summary: "Search fisherman profile"
      produces:
        - "application/json"
      parameters:
        - name: "name"
          in: "query"
          description: "Name of the fisherman to search for"
          required: false
          type: "string"
      responses:
        "200":
          description: "Search results matching the query"
        "400":
          description: "Invalid input"
  /fishermen/{id}:
    put:
      tags:
        - "Fisherman Management"
      summary: "Edit fisherman profile"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - in: "path"
          name: "id"
          description: "ID of the fisherman to edit"
          required: true
          type: "string"
        - in: "body"
          name: "body"
          description: "Fisherman object that needs to be updated"
          required: true
          schema:
            $ref: "#/definitions/Fisherman"
      responses:
        "200":
          description: "Fisherman updated"
        "400":
          description: "Invalid input"
        "404":
          description: "Fisherman not found"
    delete:
      tags:
        - "Fisherman Management"
      summary: "Delete fisherman profile"
      produces:
        - "application/json"
      parameters:
        - in: "path"
          name: "id"
          description: "ID of the fisherman to delete"
          required: true
          type: "string"
      responses:
        "204":
          description: "Fisherman deleted"
        "404":
          description: "Fisherman not found"
  /locations:
    post:
      tags:
        - "Location Management"
      summary: "Add new fishing location"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - in: "body"
          name: "body"
          description: "Location object that needs to be added to the system"
          required: true
          schema:
            $ref: "#/definitions/Location"
      responses:
        "201":
          description: "Location created"
        "400":
          description: "Invalid input"
    get:
      tags:
        - "Location Management"
      summary: "Search fishing locations"
      produces:
        - "application/json"
      parameters:
        - name: "name"
          in: "query"
          description: "Name of the fishing location to search for"
          required: false
          type: "string"
        - name: "spotType"
          in: "query"
          description: "type of fishing location"
          required: false
          type: "string"
        - name: "latitude"
          in: "query"
          description: "Latitude of the fishing location to search for"
          required: false
          type: "number"
          format: "double"
        - name: "longitude"
          in: "query"
          description: "Longitude of the fishing location to search for"
          required: false
          type: "number"
          format: "double"
      responses:
        "200":
          description: "Search results matching the query"
        "400":
          description: "Invalid input"
  /locations/{id}:
    put:
      tags:
        - "Location Management"
      summary: "Edit fishing location"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - in: "path"
          name: "id"
          description: "ID of the fishing location to edit"
          required: true
          type: "string"
        - in: "body"
          name: "body"
          description: "Location object that needs to be updated"
          required: true
          schema:
            $ref: "#/definitions/Location"
      responses:
        "200":
          description: "Location updated"
        "400":
          description: "Invalid input"
        "404":
          description: "Location not found"
    delete:
      tags:
        - "Location Management"
      summary: "Delete fishing location"
      produces:
        - "application/json"
      parameters:
        - in: "path"
          name: "id"
          description: "ID of the fishing location to delete"
          required: true
          type: "string"
      responses:
        "204":
          description: "Location deleted"
        "404":
          description: "Location not found"   
...
definitions:
  Fisherman:
    type: "object"
    required:
      - "id"
      - "name"
      - "password"
      - "age"
      - "email"
      - "phone"
      - "locationId"
    properties:
      id:
        type: "number"
      name:
        type: "string"
      password:
        type: "string"
      age:
        type: "integer"
        format: "int32"
        minimum: 18
        maximum: 99
      email:
        type: "string"
        format: "email"
      phone:
        type: "string"
  Location:
    type: "object"
    required:
      - "id"
      - "name"
      - "spotType"
      - "latitude"
      - "longitude"
      - "description"
    properties:
      id:
        type: "number"
      name:
        type: "string"
      spotType:
        type: "string"
      latitude:
        type: "number"
        format: "double"
      longitude:
        type: "number"
        format: "double"
      description:
        type: "string"  
...